refactor(consensus-rewrite-v2): move observer to lib-consensus-observer crate (closes #2330)
Implements CONS-105 per docs/epics/consensus-rewrite-epic.md and AD-003.
~3,810 LOC moved across 9 files + 1 integration test.

Closes #2330
Refs #2365
