+++ lib-consensus-observer/src/state_encoder.rs
