+++ lib-consensus/src/lib.rs
@@ -20,7 +20,6 @@ pub mod invariants;
 pub mod mempool;
 pub mod mining;
 pub mod network;
-pub mod observer;
 pub mod proofs;
 pub mod rewards;
 pub mod slashing;
